Deprecate nest_tokens
